Optimizing the plugin a little more, adding styles
