--- assets/css/truman_youtube_display.css
@@ -1,46 +1,0 @@
-/* Reference: http://avexdesigns.com/responsive-youtube-embed/ */
-
-.vid {
-	position: relative;
-	padding-bottom: 56.25%;
-	padding-top: 30px; height: 0; overflow: hidden;
-}
-
-.vid iframe,
-.vid object,
-.vid embed,
-.vid img {
-	position: absolute;
-	top: 0;
-	left: 0;
-	width: 100%;
-	height: 100%;
-}
-
-.vid span.play {
-	position: relative;
-	display: block;
-	left: 45%;
-	opacity: .7;
-	filter:alpha(opacity=70);
-	cursor: pointer;
-	border-style: solid;
-	border-width: 20px 0 20px 40px;
-	border-color: transparent transparent transparent white;
-}
-
-.vid .vid-caption {
-	position: absolute;
-	bottom: 0;
-	left: 0;
-	color: #ffffff;
-	text-align: left;
-	padding: 5px;
-	width: 100%;
-	font-weight: normal !important;
-}
-
-.vid:hover .play {
-	opacity: 1;
-	filter:alpha(opacity=100);
-}
