+++ assets/scss/wpc-youtube.scss
@@ -1,0 +1,97 @@
+.wpc-youtube {
+	overflow: auto;
+
+	*,
+	*:before,
+	*:after {
+		-webkit-box-sizing: inherit;
+		-moz-box-sizing: inherit;
+		box-sizing: inherit;
+	}
+
+	.wpc-youtube-video {
+		padding: 10px;
+		margin: 0 0 20px 0;
+
+		a {
+			display: block;
+		}
+
+		.media {
+			position: relative;
+
+			iframe,
+			object,
+			embed,
+			img {
+				width: 100%;
+				height: 100%;
+			}
+
+			img {
+				border: 1px solid #eee;
+			}
+
+			.play {
+				position: absolute;
+				left: 50%;
+				top: 50%;
+				transform: translate(-50%, -50%);
+				width: 0;
+				height: 0;
+				border-top: 35px solid transparent;
+				border-bottom: 35px solid transparent;
+				border-left: 40px solid black;
+				opacity: 0.3;
+			}
+		}
+
+		.caption {
+			display: block;
+			text-align: left;
+			padding: 10px;
+		}
+
+		&:hover {
+
+			.media {
+
+				img {
+					opacity: 0.4;
+				}
+
+				.play {
+					opacity: 1;
+				}
+			}
+		}
+	}
+
+	@media screen and (min-width: 600px) {
+
+		.wpc-youtube-video {
+			width: 50%;
+			float: left;
+
+			&:nth-child(odd) {
+				clear: both;
+			}
+		}
+	}
+
+	@media screen and (min-width: 980px) {
+
+		.wpc-youtube-video {
+			width: 33%;
+			float: left;
+
+			&:nth-child(odd) {
+				clear: none;
+			}
+
+			&:nth-of-type(3n+4) {
+				clear: both;
+			}
+		}
+	}
+}
