+++ package.json
@@ -20,6 +20,8 @@
   "homepage": "https://github.com/wpcampus/youtube-display-plugin#readme",
   "devDependencies": {
     "gulp": "^3.9.1",
-    "gulp-phpcs": "^1.4.0"
+    "gulp-autoprefixer": "^3.1.1",
+    "gulp-phpcs": "^1.4.0",
+    "gulp-sass": "^3.1.0"
   }
 }
